rename comments.tsx to commentThread.tsx
