+++ src/screens/dealInfo/comments/commentThread.tsx
